+++ src/admin-role-form/camel-case.filter.spec.js
@@ -13,31 +13,26 @@
  * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org. 
  */
 
-(function() {
+describe('camel case filter', function() {
 
-    'use strict';
+    var camelCaseFilter, $filter;
 
-    /**
-     * @module admin-user-roles
-     *
-     * @description
-     * Provides add/remove user roles screen.
-     */
-    angular.module('admin-user-roles', [
-        'mgcrea.ngStrap.tab',
-        'openlmis-modal',
-        'openlmis-pagination',
-        'openlmis-rights',
-        'referencedata-facility',
-        'referencedata-program',
-        'referencedata-role',
-        'referencedata-supervisory-node',
-        'referencedata-user',
-        'ui.router',
-        'admin-user-list',
-        'openlmis-admin',
-        'admin-role-form',
-        'openlmis-object-utils'
-    ]);
+    beforeEach(function() {
+        module('admin-role-form');
 
-})();
+        inject(function($injector) {
+            $filter = $injector.get('$filter');
+        });
+
+        camelCaseFilter = $filter('camelCase');
+    });
+
+    it('should return undefined for undefined', function() {
+        expect(camelCaseFilter()).toBeUndefined();
+    });
+
+    it('should return translated message for ', function() {
+        expect(camelCaseFilter('SOME_REALLY_LONG_NAME')).toEqual('someReallyLongName');
+    });
+
+});